First authorized api call (users)

+++ MyTodo_Todos/Controllers/TodosController.cs
@@ -17,7 +17,7 @@ namespace MyTodo_Todos.Controllers
 
         // GET api/<TodosController>/5
         [HttpGet("{id}")]
-        public string Get(int id)
+        public string Get(long id)
         {
             return "value";
         }
@@ -30,13 +30,13 @@ namespace MyTodo_Todos.Controllers
 
         // PUT api/<TodosController>/5
         [HttpPut("{id}")]
-        public void Put(int id, [FromBody] string value)
+        public void Put(long id, [FromBody] string value)
         {
         }
 
         // DELETE api/<TodosController>/5
         [HttpDelete("{id}")]
-        public void Delete(int id)
+        public void Delete(long id)
         {
         }
     }
